Restructured and reorganized code, fixed a few bugs regarding removing old links when replacing a performing a links update, encapsulated larger code blocks into functions, cleaned up comments and finished writing doc strings for all functions in modify module.
